change epoch type

+++ data/api/apiBlock.go
@@ -66,7 +66,7 @@ type ExecutionResult struct {
 	HeaderHash             string       `json:"headerHash"`
 	HeaderNonce            uint64       `json:"headerNonce"`
 	HeaderRound            uint64       `json:"headerRound"`
-	HeaderEpoch            uint64       `json:"headerEpoch"`
+	HeaderEpoch            uint32       `json:"headerEpoch"`
 	RootHash               string       `json:"rootHash"`
 	ReceiptsHash           string       `json:"receiptsHash"`
 	MiniBlocks             []*MiniBlock `json:"miniBlocks,omitempty"`
